Update README.md
The Decision Prediction Model is a machine learning solution designed to predict outcomes or decisions based on various input features. It leverages historical data to train a model that can make decisions or classifications, depending on the problem at hand. The model is highly adaptable and can be customized to predict outcomes for a wide range of decision-making scenarios, from business forecasts to medical diagnoses.